requirements.txt
# ============================================
# 项目依赖管理
# ============================================

# 核心框架
fastapi==0.110.0
uvicorn==0.29.0

# 数据校验
pydantic>=2.0.0

# 数据存储
redis==4.6.0
fakeredis==2.10.0

# 快速实体匹配
flashtext==2.7

# HTTP 客户端
requests==2.31.0

# 开发工具
python-dotenv==1.0.0

# 前端相关
# 注意：前端使用HTML5 + CSS3 + JavaScript，无需Python依赖
